+++ src/Field.php
@@ -360,7 +360,7 @@ class Field extends HtmlField
             'ui' => [
                 'viewportOffset' => ['top' => 50],
                 'poweredBy' => [
-                    'position' => 'inside',
+                    'position' => 'outside',
                     'label' => '',
                 ],
             ],